add gc content output

+++ src/lib.rs
@@ -9,15 +9,17 @@ pub fn get_n_bases(seq: &[u8]) -> i64 {
     n
 }
 // return per read content
-pub fn get_gc_content(seq: &[u8]) -> f64 {
-    let mut n: i64 = 0;
+pub fn get_gc_content(seq: &[u8]) -> f32 {
+    
+let mut n: i64 = 0;
     for s in seq {
         if s == &103u8 || s == &71u8 || s == &99u8 || s == &67u8 {
             n += 1;
         }
     }
-    n as f64 / seq.len() as f64
-}
+        n as f32 / seq.len() as f32
+    }
+
 
 pub fn get_qual_bases(q: &[u8], qx: u8) -> i64 {
     let mut n = 0;
